Fix Explorer Metadata (#2)
It doesn't look like Discord, Slack -- and possibly other services that
scrape the metadata to display rich content when links are posted
are aware of Angular's existence. For this reason, people sharing a link
to the explorer usually yields an awkward "Testnet BitcoinZ" -- even
though the explorer is running on mainnet.

I don't personally run explorers running in testnet that are public (or
if they are, they're not shared with people). I think for the time
being, it's not necessary to define which network the explorer is
connected to via metadata.